pyproject.toml
[project]
name = "broskills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "requests>=2.32.5",
    "pythainlp>=5.1.2",
    "attacut>=1.0.6",
    "pyyaml>=5.1.2",
]

[dependency-groups]
dev = [
    "rouge-score>=0.1.2",
    "ipykernel>=7.1.0",
    "pytest>=9.0.2",
]
